gradio_chatbot_test/requirements.txt
gradio>=4.44.0
chromadb>=0.5.5
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
numpy>=1.26.0
python-dotenv>=1.0.1
openai>=1.43.0
FlagEmbedding>=1.2.10      # (선택) BGE reranker 사용 시
